Changed the id value in the test - doesn't match test data.

+++ lams_common/test/java/org/lamsfoundation/lams/usermanagement/dao/hibernate/TestWorkspaceDAO.java
@@ -35,7 +35,7 @@ public class TestWorkspaceDAO extends AbstractLamsTestCase {
 	}
 	public void testGetWorkspaceByRootFolderID(){
 		boolean rootFolder = false;
-		Workspace workspace =workspaceDAO.getWorkspaceByRootFolderID(new Integer(8));
+		Workspace workspace =workspaceDAO.getWorkspaceByRootFolderID(new Integer(6));
 		if(workspace!=null)
 			rootFolder=true;
 		assertTrue(rootFolder);
